Se agregan los test de Sorpresa y se cambia el redondeo del los porcentajes por el valor exacto

+++ src/main/java/edu/fiuba/algo3/Auto.java
@@ -5,7 +5,7 @@ public class Auto extends Vehiculo {
     Auto(){
         this.movimientos = 0;
     }
-    Auto(int movimientos){
+    Auto(double movimientos){
         this.movimientos = movimientos;
     }
